# 🎯 Modular Architecture Refactor - Complete Scraper Modernization