--- a/src/content/controller.js
+++ b/src/content/controller.js
@@ -1,0 +1,10 @@
+// Main scraping workflow orchestration
+// Extracted from scraper.js during modularization
+
+// This module will contain:
+// - startScraping()
+// - stopScraping()
+// - checkContinueScraping()
+// - Main workflow coordination
+
+console.log('controller.js module loaded');