--- a/src/content/entry.js
+++ b/src/content/entry.js
@@ -1,0 +1,9 @@
+// Main entry point for modular scraper
+// Replaces scraper.js as the bootstrap file
+
+// This module will contain:
+// - Module imports and initialization
+// - DOMContentLoaded setup
+// - Page validation and startup logic
+
+console.log('entry.js module loaded');