--- a/src/content/messageBridge.js
+++ b/src/content/messageBridge.js
@@ -1,0 +1,9 @@
+// Chrome runtime message handling
+// Extracted from scraper.js during modularization
+
+// This module will contain:
+// - chrome.runtime.onMessage listener
+// - Message routing and handling
+// - Message response coordination
+
+console.log('messageBridge.js module loaded');