--- a/src/lib/extractor.js
+++ b/src/lib/extractor.js
@@ -1,0 +1,9 @@
+// Core profile extraction logic
+// Extracted from scraper.js during modularization
+
+// This module will contain:
+// - Core DOM extraction logic from scrapeCurrentPage()
+// - Profile data extraction methods
+// - DOM traversal and data gathering
+
+console.log('extractor.js module loaded');