--- a/src/lib/pagination.js
+++ b/src/lib/pagination.js
@@ -1,0 +1,11 @@
+// Pagination detection and navigation logic
+// Extracted from scraper.js during modularization
+
+// This module will contain:
+// - getTotalPages()
+// - hasNextPage()
+// - getCurrentPage()
+// - ensureNextButtonReady()
+// - navigateToNextPage()
+
+console.log('pagination.js module loaded');