--- a/src/lib/selectors.js
+++ b/src/lib/selectors.js
@@ -1,0 +1,11 @@
+// Centralized DOM selector strategies
+// Extracted from scraper.js during modularization
+
+// This module will contain:
+// - resultSelectors array
+// - profileLinkSelectors array
+// - nameSelectors array
+// - headlineSelectors array
+// - locationSelectors array
+
+console.log('selectors.js module loaded');