… CLAUDE.md
… docs/modular_refactoring_implementation.md
… docs/refactor_plan.md
… docs/regression_fixes.md
… manifest.json
… package.json
… scraper.js.backup
… src/lib/state.js
… src/lib/storageApi.js
… src/lib/utils.js
… src/lib/validator.js